This log message contains two parts: First the changes of this revision, and then the changes of revision 19908.
This revision:

New revision of "U-umlaut in Front of Nasal".
The rule is now split, again, in two rules: One WITH epenthetic vowel and
with vowel between the target vowel and UUML, and one without neither of these.
Thereby, we avoid conflict with this rule and a:0 in "Epenthetic deletion".
Also, the comments were removed.
Added examples for a couple of rules.

Revision 19908:

Added an optional IUML to the right context of "Deleting v in gv sequences" 
(the rule also was renamed), in order to cope with flúgva - flýg (both
deletion and umlaut).

Removed right context to the right of UUML in the General U-umlaut rule. and in
U-umlaut in front of nasal.
Now, the only trigger is UUML. Words like øll and lond are now ok.

Added i:e to the "eI-umlaut" rule.

Added bare AB3 as right context to Past tense singular (and plural)
 monophthongs. 
Keep and eye at those rules, for overgeneralisation.

Added á to the Vx set of "Supine i".