Errors and missing letters for hfst-tokenise detected using devtools/check_analysis_regressions.sh -dx.
When I started (before the previous fao commit) -dx produced 237 diffs, now we are down to 120. Of these, most of the diffs are real diffs - diffs in the ways dynamic compounds are handled, and how mwe entries are used. The only real error detected this way is related to the pronoun 'hasin' in various forms. I have not been able to find the cause of this error.