Moved test cases to relevant rule (lls > ls).
removed final # from himalnum test

attd ([r|l|n]) to right context of
 "Adjusting Dental Past Suffix -d(i)"
in order to allow for d:t alternation of not only
keypdi > keypti 
but also
merkdi > merkti

Adde test for it as well.

Added rule
+"stem-final i to j in front of back vowel suffix"
in order to do
merki>ar 
merkj0ar

Still open whether this should be done here or in lexc 
(the alternative is to do it with merk as stem and then add either 
i or jar)

Removed wrong test býnum:bynum.



git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/fao@182668 c7155fb1-f0a7-4240-a2fc-2600b6f42f90